+++ packages/core/src/util.object.ts
@@ -1,5 +1,11 @@
 import type {
-	Arguments, BaseFormatters, BaseTranslation, Cache, Locale, TranslationFunctions, TranslationKey
+	Arguments,
+	BaseFormatters,
+	BaseTranslation,
+	Cache,
+	Locale,
+	TranslationFunctions,
+	TranslationKey,
 } from './core'
 import { translate } from './core'
 import { getPartsFromString } from './util.string'
@@ -7,7 +13,7 @@ import { getPartsFromString } from './util.string'
 type TranslateByKey<T extends BaseTranslation> = (key: TranslationKey<T>, ...args: Arguments) => string
 
 const getTextFromTranslationKey = <T extends BaseTranslation>(translations: T, key: TranslationKey<T>): string => {
-	; (key as string).split('.').forEach((k) => (translations = translations[k] as T))
+	;(key as string).split('.').forEach((k) => (translations = translations[k] as T))
 	return (translations as unknown as string) ?? (key as string)
 }
 
@@ -27,20 +33,25 @@ export function i18nObject<
 	T extends BaseTranslation,
 	TF extends TranslationFunctions = TranslationFunctions<T>,
 	F extends BaseFormatters = BaseFormatters,
-	>(locale: L, translations: T, formatters: F): TF
+>(locale: L, translations: T, formatters: F): TF
 
 export function i18nObject<
 	L extends Locale,
 	T extends BaseTranslation,
 	TF extends TranslationFunctions = TranslationFunctions<T>,
-	>(locale: L, translations: T): TF
+>(locale: L, translations: T): TF
 
 // eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types, @typescript-eslint/no-explicit-any
 export function i18nObject(locale: any, translations: any, formatters: any = {}): any {
 	return createProxy(getTranslateInstance(locale, translations, formatters))
 }
 
-const createProxy = <T extends BaseTranslation>(fn: TranslateByKey<T>, prefixKey?: string, proxyObject = {}): TranslationFunctions<T> =>
+const createProxy = <T extends BaseTranslation>(
+	fn: TranslateByKey<T>,
+	prefixKey?: string,
+	proxyObject = {},
+): TranslationFunctions<T> =>
 	new Proxy((prefixKey ? fn.bind(null, prefixKey) : proxyObject) as TranslationFunctions<T>, {
-		get: (target, key: string) => !(target === proxyObject && key === 'then') && createProxy(fn, prefixKey ? `${prefixKey}.${key}` : key),
+		get: (target, key: string) =>
+			!(target === proxyObject && key === 'then') && createProxy(fn, prefixKey ? `${prefixKey}.${key}` : key),
 	})