+++ packages/generator/src/files/generate-adapter-svelte.ts
@@ -1,21 +1,38 @@
 import { writeFileIfContainsChanges } from '../file-utils'
 import { GeneratorConfigWithDefaultValues } from '../generate-files'
 import { prettify } from '../generator-util'
-import { generics, importTypes, jsDocFunction, jsDocImports, jsDocType, OVERRIDE_WARNING, tsCheck, type } from '../output-handler'
+import {
+	generics,
+	importTypes,
+	jsDocFunction,
+	jsDocImports,
+	jsDocType,
+	OVERRIDE_WARNING,
+	tsCheck,
+	type,
+} from '../output-handler'
 
-const getSvelteUtils = (
-	{ baseLocale, formattersTemplateFileName, typesFileName, utilFileName, banner }: GeneratorConfigWithDefaultValues,
-) => {
+const getSvelteUtils = ({
+	baseLocale,
+	formattersTemplateFileName,
+	typesFileName,
+	utilFileName,
+	banner,
+}: GeneratorConfigWithDefaultValues) => {
 	return `${OVERRIDE_WARNING}${tsCheck}
 ${banner}
 
 ${jsDocImports(
-		{ from: 'typesafe-i18n/adapters/adapter-svelte', type: 'SvelteStoreInit<Locales, Translation, TranslationFunctions>', alias: 'SvelteStoreInit' },
-		{ from: `./${typesFileName}`, type: 'Locales' },
-		{ from: `./${typesFileName}`, type: 'Translation' },
-		{ from: `./${typesFileName}`, type: 'TranslationFunctions' },
-		{ from: `./${typesFileName}`, type: 'Formatters' },
-	)}
+	{
+		from: 'typesafe-i18n/adapters/adapter-svelte',
+		type: 'SvelteStoreInit<Locales, Translation, TranslationFunctions>',
+		alias: 'SvelteStoreInit',
+	},
+	{ from: `./${typesFileName}`, type: 'Locales' },
+	{ from: `./${typesFileName}`, type: 'Translation' },
+	{ from: `./${typesFileName}`, type: 'TranslationFunctions' },
+	{ from: `./${typesFileName}`, type: 'Formatters' },
+)}
 
 import { getI18nSvelteStore } from 'typesafe-i18n/adapters/adapter-svelte';
 ${importTypes(`./${typesFileName}`, 'Locales', 'Translation', 'TranslationFunctions', 'Formatters')}
@@ -23,7 +40,12 @@ import { getTranslationForLocale } from './${utilFileName}'
 import { initFormatters } from './${formattersTemplateFileName}'
 
 ${jsDocType('SvelteStoreInit')}
-const { initI18n: init, setLocale, isLoadingLocale, locale, LL } = getI18nSvelteStore${generics('Locales', 'Translation', 'TranslationFunctions', 'Formatters')}()
+const { initI18n: init, setLocale, isLoadingLocale, locale, LL } = getI18nSvelteStore${generics(
+		'Locales',
+		'Translation',
+		'TranslationFunctions',
+		'Formatters',
+	)}()
 
 ${jsDocFunction('Promise<void>', { type: 'Locales', name: 'locale' })}
 const initI18n = (locale${type('Locales')} = '${baseLocale}') => init(locale, getTranslationForLocale, initFormatters)
@@ -34,9 +56,7 @@ export default LL
 `
 }
 
-export const generateSvelteAdapter = async (
-	config: GeneratorConfigWithDefaultValues,
-): Promise<void> => {
+export const generateSvelteAdapter = async (config: GeneratorConfigWithDefaultValues): Promise<void> => {
 	const { outputPath } = config
 
 	const svelteUtils = getSvelteUtils(config)