+++ packages/generator/src/files/generate-template-formatters.ts
@@ -3,24 +3,23 @@ import { GeneratorConfigWithDefaultValues } from '../generate-files'
 import { prettify } from '../generator-util'
 import { importTypes, jsDocFunction, jsDocImports, jsDocType, tsCheck, type } from '../output-handler'
 
-const getFormattersTemplate = (
-	{ typesFileName: typesFile, loadLocalesAsync }: GeneratorConfigWithDefaultValues,
-) => {
+const getFormattersTemplate = ({ typesFileName: typesFile, loadLocalesAsync }: GeneratorConfigWithDefaultValues) => {
 	const formattersInitializerType = `${loadLocalesAsync ? 'Async' : ''}FormattersInitializer`
 	return `${tsCheck}
 
 ${jsDocImports(
-		{ from: 'typesafe-i18n', type: 'FormattersInitializer<Locales, Formatters>', alias: 'FormattersInitializer' },
-		{ from: `./${typesFile}`, type: 'Locales' },
-		{ from: `./${typesFile}`, type: 'Formatters' },
-	)}
+	{ from: 'typesafe-i18n', type: 'FormattersInitializer<Locales, Formatters>', alias: 'FormattersInitializer' },
+	{ from: `./${typesFile}`, type: 'Locales' },
+	{ from: `./${typesFile}`, type: 'Formatters' },
+)}
 
 ${importTypes('typesafe-i18n', formattersInitializerType)}
 ${importTypes(`./${typesFile}`, 'Locales', 'Formatters')}
 
 ${jsDocFunction(loadLocalesAsync ? 'Promise<Formatters>' : 'Formatters', { type: 'Locales', name: 'locale' })}
-export const initFormatters${type(`${formattersInitializerType}<Locales, Formatters>`)} = ${loadLocalesAsync ? 'async ' : ''
-		}(locale${type('Locales')}) => {
+export const initFormatters${type(`${formattersInitializerType}<Locales, Formatters>`)} = ${
+		loadLocalesAsync ? 'async ' : ''
+	}(locale${type('Locales')}) => {
 	${jsDocType('Formatters')}
 	const formatters${type('Formatters')} = {
 		// add your formatter functions here