+++ packages/generator/src/files/generate-util.ts
@@ -2,7 +2,20 @@ import type { Locale } from '../../../core/src/core'
 import { writeFileIfContainsChanges } from '../file-utils'
 import { GeneratorConfigWithDefaultValues } from '../generate-files'
 import { prettify, sanitizeLocale } from '../generator-util'
-import { generics, importTypes, jsDocFunction, jsDocImports, jsDocTsIgnore, jsDocType, OVERRIDE_WARNING, tsCheck, type, typeCast } from '../output-handler'
+import {
+	generics,
+	importTypes,
+	jsDocFunction,
+	jsDocImports,
+	jsDocTsIgnore,
+	jsDocType,
+	OVERRIDE_WARNING,
+	relativeFileImportPath,
+	relativeFolderImportPath,
+	tsCheck,
+	type,
+	typeCast
+} from '../output-handler'
 
 const getLocalesTranslationRowAsync = (locale: Locale): string => {
 	const sanitizedLocale = sanitizeLocale(locale)
@@ -11,7 +24,7 @@ const getLocalesTranslationRowAsync = (locale: Locale): string => {
 	const wrappedLocale = needsEscaping ? `'${locale}'` : locale
 
 	return `
-	${wrappedLocale}: () => import('./${locale}'),`
+	${wrappedLocale}: () => import('${relativeFolderImportPath(locale)}'),`
 }
 
 const getAsyncCode = ({ locales }: GeneratorConfigWithDefaultValues) => {
@@ -23,10 +36,19 @@ const localeTranslationLoaders = {${localesTranslationLoaders}
 }
 
 ${jsDocFunction('Promise<Translation>', { type: 'Locales', name: 'locale' })}
-export const getTranslationForLocale = async (locale${type('Locales')}) => (await (localeTranslationLoaders[locale] || localeTranslationLoaders[baseLocale])()).default${typeCast('Translation')}
+export const getTranslationForLocale = async (locale${type(
+		'Locales',
+	)}) => (await (localeTranslationLoaders[locale] || localeTranslationLoaders[baseLocale])()).default${typeCast(
+		'Translation',
+	)}
 
 ${jsDocFunction('Promise<TranslationFunctions>', { type: 'Locales', name: 'locale' })}
-export const i18nObject = (locale${type('Locales')}) => i18nObjectLoaderAsync${generics('Locales', 'Translation', 'TranslationFunctions', 'Formatters')}(locale, getTranslationForLocale, initFormatters)
+export const i18nObject = (locale${type('Locales')}) => i18nObjectLoaderAsync${generics(
+		'Locales',
+		'Translation',
+		'TranslationFunctions',
+		'Formatters',
+	)}(locale, getTranslationForLocale, initFormatters)
 `
 }
 
@@ -35,7 +57,11 @@ const getLocalesTranslationRowSync = (locale: Locale, baseLocale: string): strin
 	const needsEscaping = locale !== sanitizedLocale
 
 	const postfix =
-		locale === baseLocale ? `: ${sanitizedLocale}${typeCast('Translation')}` : needsEscaping ? `: ${sanitizedLocale}` : ''
+		locale === baseLocale
+			? `: ${sanitizedLocale}${typeCast('Translation')}`
+			: needsEscaping
+			? `: ${sanitizedLocale}`
+			: ''
 
 	const wrappedLocale = needsEscaping ? `'${locale}'` : locale
 
@@ -47,7 +73,7 @@ const getSyncCode = ({ baseLocale, locales }: GeneratorConfigWithDefaultValues) 
 	const localesImports = locales
 		.map(
 			(locale) => `
-import ${sanitizeLocale(locale)} from './${locale}'`,
+import ${sanitizeLocale(locale)} from '${relativeFolderImportPath(locale)}'`,
 		)
 		.join('')
 
@@ -59,13 +85,25 @@ const localeTranslations${type('LocaleTranslations<Locales, Translation>')} = {$
 }
 
 ${jsDocFunction('Translation', { type: 'Locales', name: 'locale' })}
-export const getTranslationForLocale = (locale${type('Locales')}) => localeTranslations[locale] || localeTranslations[baseLocale]
+export const getTranslationForLocale = (locale${type(
+		'Locales',
+	)}) => localeTranslations[locale] || localeTranslations[baseLocale]
 
 ${jsDocFunction('TranslationFunctions', { type: 'Locales', name: 'locale' })}
-export const i18nObject = (locale${type('Locales')}) => i18nObjectLoader${generics('Locales', 'Translation', 'TranslationFunctions', 'Formatters')}(locale, getTranslationForLocale, initFormatters)
+export const i18nObject = (locale${type('Locales')}) => i18nObjectLoader${generics(
+		'Locales',
+		'Translation',
+		'TranslationFunctions',
+		'Formatters',
+	)}(locale, getTranslationForLocale, initFormatters)
 
 ${jsDocFunction('LocaleTranslationFunctions')}
-export const i18n = () => initI18n${generics('Locales', 'Translation', 'TranslationFunctions', 'Formatters')}(getTranslationForLocale, initFormatters)
+export const i18n = () => initI18n${generics(
+		'Locales',
+		'Translation',
+		'TranslationFunctions',
+		'Formatters',
+	)}(getTranslationForLocale, initFormatters)
 `
 }
 
@@ -76,7 +114,7 @@ const getUtil = (config: GeneratorConfigWithDefaultValues): string => {
 		loadLocalesAsync,
 		baseLocale,
 		locales,
-		banner
+		banner,
 	} = config
 
 	const dynamicImports = loadLocalesAsync
@@ -98,20 +136,20 @@ export const locales${type('Locales[]')} = [${locales.map(
 ${banner}
 
 ${jsDocImports(
-		{ from: 'typesafe-i18n', type: 'TranslateByString' },
-		{ from: 'typesafe-i18n', type: 'LocaleTranslations<Locales, Translation>', alias: 'LocaleTranslations' },
-		!loadLocalesAsync ? { from: 'typesafe-i18n', type: 'LocaleTranslationFunctions' } : undefined,
-		{ from: 'typesafe-i18n/detectors', type: 'LocaleDetector' },
-		{ from: `./${typesFileName}`, type: 'Locales' },
-		{ from: `./${typesFileName}`, type: 'Translation' },
-		{ from: `./${typesFileName}`, type: 'TranslationFunctions' },
-	)}
+	{ from: 'typesafe-i18n', type: 'TranslateByString' },
+	{ from: 'typesafe-i18n', type: 'LocaleTranslations<Locales, Translation>', alias: 'LocaleTranslations' },
+	!loadLocalesAsync ? { from: 'typesafe-i18n', type: 'LocaleTranslationFunctions' } : undefined,
+	{ from: 'typesafe-i18n/detectors', type: 'LocaleDetector' },
+	{ from: `./${typesFileName}`, type: 'Locales' },
+	{ from: `./${typesFileName}`, type: 'Translation' },
+	{ from: `./${typesFileName}`, type: 'TranslationFunctions' },
+)}
 
 ${dynamicImports}
 ${importTypes(`./${typesFileName}`, 'Translation', 'TranslationFunctions', 'Formatters', 'Locales')}
 ${importTypes('typesafe-i18n/detectors', 'LocaleDetector')}
 import { detectLocale as detectLocaleFn } from 'typesafe-i18n/detectors'
-import { initFormatters } from './${formattersTemplatePath}'
+import { initFormatters } from '${relativeFileImportPath(formattersTemplatePath)}'
 
 ${jsDocType('Locales')}
 export const baseLocale${type('Locales')} = '${baseLocale}'
@@ -119,13 +157,19 @@ export const baseLocale${type('Locales')} = '${baseLocale}'
 ${localesEnum}
 ${dynamicCode}
 
-${jsDocFunction(loadLocalesAsync ? 'Promise<TranslateByString>' : 'TranslateByString', { type: 'Locales', name: 'locale' })}
-export const i18nString = ${loadLocalesAsync ? 'async ' : ''
-		}(locale${type('Locales')}) => initI18nString${generics('Locales', 'Formatters')}(locale, ${loadLocalesAsync ? 'await ' : ''
-		}initFormatters(locale))
+${jsDocFunction(loadLocalesAsync ? 'Promise<TranslateByString>' : 'TranslateByString', {
+	type: 'Locales',
+	name: 'locale',
+})}
+export const i18nString = ${loadLocalesAsync ? 'async ' : ''}(locale${type('Locales')}) => initI18nString${generics(
+		'Locales',
+		'Formatters',
+	)}(locale, ${loadLocalesAsync ? 'await ' : ''}initFormatters(locale))
 
 ${jsDocFunction('Locales', { type: 'LocaleDetector[]', name: 'detectors' })}
-export const detectLocale = (...detectors${type('LocaleDetector[]')}) => detectLocaleFn${generics('Locales')}(baseLocale, locales, ...detectors)
+export const detectLocale = (...detectors${type('LocaleDetector[]')}) => detectLocaleFn${generics(
+		'Locales',
+	)}(baseLocale, locales, ...detectors)
 `
 }