+++ packages/generator/src/generator.ts
@@ -5,7 +5,6 @@ import * as ts from 'typescript'
 import type { BaseTranslation } from '../../core/src/core'
 import {
 	containsFolders,
-	copyFile,
 	createPathIfNotExits,
 	deleteFolderRecursive,
 	doesPathExist,
@@ -13,21 +12,32 @@ import {
 	getFiles,
 	importFile
 } from './file-utils'
-import { generate, GeneratorConfig, GeneratorConfigWithDefaultValues, getConfigWithDefaultValues, readConfig } from './generate-files'
-import { logger, parseTypescriptVersion, TypescriptVersion } from './generator-util'
+import {
+	generate,
+	GeneratorConfig,
+	GeneratorConfigWithDefaultValues,
+	getConfigWithDefaultValues,
+	readConfig
+} from './generate-files'
+import { createLogger, Logger, parseTypescriptVersion, TypescriptVersion } from './generator-util'
 import { configureOutputHandler, fileEnding, shouldGenerateJsDoc } from './output-handler'
+
+let logger: Logger
 
 const getAllLanguages = async (path: string) => {
 	const files = await getFiles(path, 1)
 	return files.filter(({ folder, name }) => folder && name === `index${fileEnding}`).map(({ folder }) => folder)
 }
 
-
 /**
  * looks for the location of the compiled 'index.js' file
  * if the 'index.ts' file imports something from outside it's directory, we need to find the correct path to the base location file
  */
-const detectLocationOfCompiledBaseTranslation = async (outputPath: string, locale: string, tempPath: string): Promise<string> => {
+const detectLocationOfCompiledBaseTranslation = async (
+	outputPath: string,
+	locale: string,
+	tempPath: string,
+): Promise<string> => {
 	if (!containsFolders(tempPath)) return ''
 
 	const directory = await getDirectoryStructure(tempPath)
@@ -47,7 +57,9 @@ const detectLocationOfCompiledBaseTranslation = async (outputPath: string, local
 			while (isPathValid && outputPathPartsRest.length) {
 				// we need to find the full matching path
 				// e.g. `src/path/i18n/en` is invalid if the base locale is located inside `src/i18n/en`
-				const subSubDirectoryOfCurrentSection = subDirectoryOfCurrentSection[outputPathPartsRest[0] as string] as Record<string, unknown>
+				const subSubDirectoryOfCurrentSection = subDirectoryOfCurrentSection[
+					outputPathPartsRest[0] as string
+				] as Record<string, unknown>
 				if (subSubDirectoryOfCurrentSection) {
 					subPaths.push(outputPathPartsRest[0] as string)
 					outputPathPartsRest = outputPathPartsRest.slice(1)
@@ -67,25 +79,18 @@ const detectLocationOfCompiledBaseTranslation = async (outputPath: string, local
 	return ''
 }
 
-const transpileTypescriptAndPrepareImportFile = async (outputPath: string, languageFilePath: string, locale: string, tempPath: string): Promise<string> => {
+const transpileTypescriptFiles = async (
+	outputPath: string,
+	languageFilePath: string,
+	locale: string,
+	tempPath: string,
+): Promise<string> => {
 	const program = ts.createProgram([languageFilePath], { outDir: tempPath })
 	program.emit()
 
 	const baseTranslationPath = await detectLocationOfCompiledBaseTranslation(outputPath, locale, tempPath)
 
-	const compiledPath = resolve(tempPath, `${baseTranslationPath}index.js`)
-	const copyPath = resolve(tempPath, `${baseTranslationPath}i18n-temp-${debounceCounter}.js`)
-
-	const copySuccess = await copyFile(compiledPath, copyPath, false)
-	if (!copySuccess) {
-		logger.error(
-			`Make sure to give your base locale's default export the type of 'BaseTranslation' and to name the file 'index.ts'.
-See https://github.com/ivanhofer/typesafe-i18n#folder-structure for more info`,
-		)
-		return ''
-	}
-
-	return copyPath
+	return resolve(tempPath, `${baseTranslationPath}index.js`)
 }
 
 const parseLanguageFile = async (
@@ -100,11 +105,11 @@ const parseLanguageFile = async (
 		return null
 	}
 
-	!shouldGenerateJsDoc && await createPathIfNotExits(tempPath)
+	!shouldGenerateJsDoc && (await createPathIfNotExits(tempPath))
 
 	const importPath = shouldGenerateJsDoc
 		? originalPath
-		: await transpileTypescriptAndPrepareImportFile(outputPath, originalPath, locale, tempPath)
+		: await transpileTypescriptFiles(outputPath, originalPath, locale, tempPath)
 
 	if (!importPath) {
 		return null
@@ -112,7 +117,7 @@ const parseLanguageFile = async (
 
 	const languageImport = await importFile<BaseTranslation>(importPath)
 
-	!shouldGenerateJsDoc && await deleteFolderRecursive(tempPath)
+	!shouldGenerateJsDoc && (await deleteFolderRecursive(tempPath))
 
 	if (!languageImport) {
 		logger.error(`could not read default export from language file '${locale}'`)
@@ -136,9 +141,7 @@ const parseAndGenerate = async (config: GeneratorConfigWithDefaultValues, versio
 	const locales = await getAllLanguages(outputPath)
 
 	const languageFile =
-		(locale && (await parseLanguageFile(outputPath, locale, tempPath))) || {}
-
-	if (languageFile) return
+		(locale && (await parseLanguageFile(outputPath, locale, resolve(tempPath, `${debounceCounter}`)))) || {}
 
 	await generate(languageFile, { ...config, baseLocale: locale, locales }, version, logger)
 
@@ -157,6 +160,8 @@ const debounce = (callback: () => void) =>
 	)
 
 export const startGenerator = async (config?: GeneratorConfig, watchFiles = true): Promise<void> => {
+	logger = createLogger(console, !watchFiles)
+
 	const configWithDefaultValues = await getConfigWithDefaultValues(config)
 	const { outputPath } = configWithDefaultValues
 
@@ -169,9 +174,10 @@ export const startGenerator = async (config?: GeneratorConfig, watchFiles = true
 
 	watchFiles && watch(outputPath).on('all', () => debounce(onChange))
 
-	logger.info(`generating files for ${shouldGenerateJsDoc
-		? 'JavaScript with JSDoc notation'
-		: `TypeScript version: '${ts.versionMajorMinor}.x'`}`
+	logger.info(
+		`generating files for ${
+			shouldGenerateJsDoc ? 'JavaScript with JSDoc notation' : `TypeScript version: '${ts.versionMajorMinor}.x'`
+		}`,
 	)
 	logger.info(`options:`, await readConfig(config))
 	watchFiles && logger.info(`watcher started in: '${outputPath}'`)