+++ packages/node-process/src/node-generator.ts
@@ -1,5 +1,5 @@
-import { logger } from 'packages/generator/src/generator-util';
-import { startGenerator } from '../../generator/src/generator';
+import { logger } from 'packages/generator/src/generator-util'
+import { startGenerator } from '../../generator/src/generator'
 
 type AllowedArguments = '--no-watch'
 
@@ -7,10 +7,10 @@ const allowedArguments: AllowedArguments[] = ['--no-watch']
 
 const [, , ...args] = process.argv as [string, string, ...AllowedArguments[]]
 
-args.forEach(arg => {
+args.forEach((arg) => {
 	if (!allowedArguments.includes(arg)) {
 		logger.error(`unknown argument '${arg}'`)
-		logger.info(`allowed arguments: ${allowedArguments.map(arg => `'${arg}'`).join(', ')}`)
+		logger.info(`allowed arguments: ${allowedArguments.map((arg) => `'${arg}'`).join(', ')}`)
 		process.exit(1)
 	}
 })